client/systray: move cmd/systray to client/systray
Updates #1708
